Update Gruntfile.js
Fixes a bug that happens on windows causing grunt to fail at all tasks... not sure if this is worthy of a PR since it might only apply to windows, but figured I'd toss it out there anyway. This fixed my and another windows users issue, and *doesn't* stop it from working on my Mac.